pyproject.toml
[project]
name = "terminusgps-payments"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authorizenet>=1.1.6",
    "django>=5.2.6",
    "python-terminusgps>=42.2.4",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["terminusgps_payments/"]

[tool.pyright]
strictListInference = false
strictDictionaryInference = false
strictSetInference = false
typeCheckingMode = "basic"

[tool.ruff]
exclude = [
    ".git",
    ".gitignore",
    ".venv",
    "venv",
    ".vscode",
    "__pycache__",
    "site-packages",
    "dist",
    "build",
]
line-length = 79
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
]
ignore = ["F821"]
fixable = [ "ALL" ]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = false
skip-magic-trailing-comma = true
